feat: threads are all good
