--- hexBoy/db/query/createDatabase.py
@@ -1,4 +1,0 @@
-# this doesn't work
-
-from hexBoy.db.logger.HexDBSetup import createDatabase
-createDatabase();
