+++ hexBoy/db/query/listGames.py
@@ -1,11 +1,12 @@
 from sqlalchemy import select
 from sqlalchemy.orm import Session
-from hexBoy.db.logger.HexDBSetup import HexLogger, Game
+from hexBoy.db.HexDBConfig import  Game, Move
+from hexBoy.db.HexQuery import  HexQuery
 
-def listGames(param: str) -> None:
+def listGames() -> None:
     # Print out the list of all of the games
-    xLogger = HexLogger()
-    with Session(xLogger.engine) as session:
+    xQ = HexQuery()
+    with Session(xQ.engine) as session:
         query = (
             select(Game)
         )
