+++ hexBoy/db/query/listMovesForGame.py
@@ -1,13 +1,14 @@
 from sqlalchemy import select
 from sqlalchemy.orm import Session
-from hexBoy.db.logger.HexDBSetup import HexLogger, Game, Move
+from hexBoy.db.HexDBConfig import  Game, Move
+from hexBoy.db.HexQuery import  HexQuery
 
 def listMovesForGame(param: str) -> None:
     """List all of the moves for a given game."""
     gameId = int(param)
 
-    xLogger = HexLogger()
-    with Session(xLogger.engine) as session:
+    xQ = HexQuery()
+    with Session(xQ.engine) as session:
         query = (
             select(Move)
             .where(Move.game_id == gameId)
