+++ hexBoy/db/query/listMovesForLastGame.py
@@ -1,13 +1,15 @@
 from sqlalchemy import select, func
 from sqlalchemy.orm import Session
-from hexBoy.db.logger.HexDBSetup import HexLogger, Game, Move
+from hexBoy.db.HexDBConfig import  Game, Move
+from hexBoy.db.HexQuery import  HexQuery
+
 
 def listMovesForLastGame() -> None:
     """List all of the  moves for the last game played."""
 
     # Print out the list of all of the games
-    xLogger = HexLogger()
-    with Session(xLogger.engine) as session:
+    xQ = HexQuery()
+    with Session(xQ.engine) as session:
         query = (
             func.max(Game.id)
         )
@@ -15,7 +17,6 @@ def listMovesForLastGame() -> None:
         # get the max game id
         gameId = session.scalar(query)
 
-            
         query = (
             select(Move)
             .where(Move.game_id == gameId)
