+++ xQuery.py
@@ -1,5 +1,5 @@
 import sys
-from hexBoy.db.logger.HexDBSetup import resetDatabase, initDB
+from hexBoy.db.HexDBConfig import resetDatabase, initDB
 from hexBoy.db.query.listGames import listGames
 from hexBoy.db.query.listMovesForGame import listMovesForGame
 from hexBoy.db.query.listMovesForLastGame import listMovesForLastGame
